Group chat displayed

+++ WebApplication2/Models/Group.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace WebApplication2.Models
 {
@@ -10,7 +12,12 @@ namespace WebApplication2.Models
         public long GroupId { get; set; }
         public string Name { get; set; }
         public List<Participant> Participants { get; set; }
-
+        public long AdminId { get; set; }
+        [ForeignKey("AdminId")]
+        public Chatter Admin { get; set; }
+        public DateTime CreatedAt { get; set; }
+        [NotMapped]
+        public string CreatedAtString { get; set; }
 
 
         public Message<object> Create()
@@ -23,10 +30,7 @@ namespace WebApplication2.Models
                     {
                         return new Message<object>(false, "Provide a valid name for group", "Group > Create", null);
                     }
-                    if (this.Participants.Count < 2)
-                    {
-                        return new Message<object>(false, "Add atleast two participants", "Group > Create", null);
-                    }
+
                     else
                     {
                         List<Participant> parties = new List<Participant>();
@@ -38,9 +42,19 @@ namespace WebApplication2.Models
                                 party.ChatterId = chatter.ChatterId;
                                 parties.Add(party);
                             }
-                            
+
                         }
+                        parties.Add(new Participant()
+                        {
+                            ChatterId = this.AdminId,
+                            Email = db.Chatters.Find(this.AdminId).Email
+                        });
                         this.Participants = parties;
+                        this.CreatedAt = DateTime.UtcNow;
+                        if (this.Participants.Count < 3)
+                        {
+                            return new Message<object>(false, "Add atleast two valid participants", "Group > Create", null);
+                        }
                         db.Groups.Add(this);
                         db.SaveChanges();
                         return new Message<object>(true, "Group created", "Group > Create", null);
@@ -64,10 +78,7 @@ namespace WebApplication2.Models
                     {
                         return new Message<object>(false, "Provide a valid name for group", "Group > Update", null);
                     }
-                    if (this.Participants.Count < 2)
-                    {
-                        return new Message<object>(false, "Add atleast two participants", "Group > Update", null);
-                    }
+
                     else
                     {
                         var group = db.Groups.Find(this.GroupId);
@@ -87,7 +98,16 @@ namespace WebApplication2.Models
                             }
 
                         }
+                        parties.Add(new Participant()
+                        {
+                            ParticipantId = this.AdminId,
+                            Email = db.Chatters.Find(this.AdminId).Email
+                        });
                         group.Participants = parties;
+                        if (this.Participants.Count < 3)
+                        {
+                            return new Message<object>(false, "Add atleast two participants", "Group > Update", null);
+                        }
                         db.Participants.RemoveRange(db.Participants.Where(x => x.GroupId == group.GroupId));
                         db.SaveChanges();
                         db.Entry(group).State = System.Data.Entity.EntityState.Modified;
@@ -125,6 +145,34 @@ namespace WebApplication2.Models
             catch (Exception ex)
             {
                 return new Message<object>(false, "Something went wrong", "Group > Create", ex);
+            }
+        }
+
+        public static Message<List<Group>> Get(long me)
+        {
+            try
+            {
+                using (Data db = new Data())
+                {
+                    var groups = db.Participants.Where(x => x.ChatterId == me).Select(x => x.Group).ToList();
+                    groups.ForEach(x =>
+                    {
+                        x.Admin = db.Chatters.Find(x.AdminId);
+                        x.CreatedAtString = x.CreatedAt.ToString("dd MMM yy");
+                    });
+
+                    if (groups != null && groups.Count != 0)
+                    {
+                        return new Message<List<Group>>(true, "Groups retrieved successfully", "Groups > Get", groups);
+                    }
+                    return new Message<List<Group>>(false, "retrieval failed");
+                }
+
+            }
+            catch (Exception ex)
+            {
+                return new Message<List<Group>>(false, "Something went wrong", "Group > Get", ex);
+                throw;
             }
         }
     }
